# frontend/openapi/schema.json
{
  "components" : {
    "schemas" : {
...
      "ApiError" : {
        "properties" : {
          "code" : {
            "type" : "string"
          },
          "detail" : {
            "type" : "string"
          },
          "field" : {
            "type" : "string"
          },
          "message" : {
            "type" : "string"
          },
          "params" : {
            "additionalProperties" : { },
            "type" : "object"
          }
        },
        "type" : "object"
      },
...
      "ApiResponseTokenResponse" : {
        "properties" : {
          "data" : {
            "$ref" : "#/components/schemas/TokenResponse"
          },
          "error" : {
            "$ref" : "#/components/schemas/ApiError"
          },
          "requestId" : {
            "type" : "string"
          }
        },
        "type" : "object"
      },
...
      "ApiResponseVoid" : {
        "properties" : {
          "data" : { },
          "error" : {
            "$ref" : "#/components/schemas/ApiError"
          },
          "requestId" : {
            "type" : "string"
          }
        },
        "type" : "object"
      },
...
      "LoginRequest" : {
        "properties" : {
          "loginId" : {
            "minLength" : 1,
            "type" : "string"
          },
          "password" : {
            "minLength" : 1,
            "type" : "string"
          }
        },
        "required" : [ "loginId", "password" ],
        "type" : "object"
      },
...
      "TokenResponse" : {
        "properties" : {
          "accessToken" : {
            "type" : "string"
          },
          "expiresIn" : {
            "format" : "int64",
            "type" : "integer"
          },
          "tokenType" : {
            "type" : "string"
          }
        },
        "type" : "object"
      },
...
    },
    "securitySchemes" : {
      "bearerAuth" : {
        "bearerFormat" : "JWT",
        "scheme" : "bearer",
        "type" : "http"
      }
    }
  },
  "openapi" : "3.1.0",
  "paths" : {
...
    "/api/v1/auth/login" : {
      "post" : {
        "operationId" : "login",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/LoginRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponseTokenResponse"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "auth-token-controller" ]
      }
    },
    "/api/v1/auth/logout" : {
      "post" : {
        "operationId" : "logout",
        "parameters" : [ {
          "in" : "cookie",
          "name" : "fgc_refresh",
          "required" : false,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponseVoid"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "auth-token-controller" ]
      }
    },
...
    "/api/v1/auth/refresh" : {
      "post" : {
        "operationId" : "refresh",
        "parameters" : [ {
          "in" : "cookie",
          "name" : "fgc_refresh",
          "required" : false,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponseTokenResponse"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "auth-token-controller" ]
      }
    },
...
  },
  "security" : [ {
    "bearerAuth" : [ ]
  } ],
  "servers" : [ {
    "url" : "/"
  } ],
...
}